a lot of everything

+++ site.sellingcoupons/lib/coupon_seller.php
@@ -1,14 +1,81 @@
 <?php
 
+/**
+ * 
+ * TODO: Переименовать класс в SoldCouoponManager
+ * TODO: Переименовать deleteCoupons => deleteSoldCoupon
+ * TODO: Переименовать sellCouopn => markCouponAsSold
+ * 
+ */
 namespace Site\SellingCoupons;
 
+use Bitrix\Main\Loader;
+use \Bitrix\Sale\Internals;
 use Site\SellingCoupons\DataMappers\SoldCouponsTable;
 
 class CouponSeller
 {
-    public function doSellCoupon()
+    /**
+     * @return \Site\SellingCoupons\DataMappers\SoldCoupon[]
+     */
+    public function sellCoupons(int $discountId, int $count): array
     {
-        throw new \Exception('TODO: incoplite method');
+        if (!Loader::includeModule('sale'))
+        {
+            return [];
+        }
+        
+        Internals\DiscountCouponTable::setDiscountCheckList([$discountId]);
+        Internals\DiscountCouponTable::disableCheckCouponsUse();
+        
+        $connection = \Bitrix\Main\Application::getConnection();
+
+        $connection->startTransaction();
+        
+        $couponClassName = Internals\DiscountCouponTable::getObjectClass();
+
+        $success = true;
+        $soldCouons = [];
+        for ($counter = 0; $counter < $count; $counter++)
+        {
+            $couponCode = Internals\DiscountCouponTable::generateCoupon(true);
+
+            $newCoupon = new $couponClassName();
+            $newCoupon->setDiscountId($discountId);
+            $newCoupon->setCoupon($couponCode);
+            $newCoupon->setType(Internals\DiscountCouponTable::TYPE_ONE_ORDER);
+            $result = $newCoupon->save();
+
+            if (!$result->isSuccess())
+            {
+                $success = false;
+                $connection->rollbackTransaction();
+                break;
+            }
+
+            $newSoldCoupon = new \Site\SellingCoupons\DataMappers\SoldCoupon();
+            $newSoldCoupon->setCoupon($newCoupon);
+            $result = $newSoldCoupon->save();
+
+            if (!$result->isSuccess())
+            {
+                $success = false;
+                $connection->rollbackTransaction();
+                break;
+            }
+
+            $soldCouons[] = $newSoldCoupon;
+        }
+
+        if ($success)
+        {
+            $connection->commitTransaction();
+        }
+
+        Internals\DiscountCouponTable::enableCheckCouponsUse();
+        Internals\DiscountCouponTable::updateUseCoupons();
+       
+        return $success ? $soldCouons : [];
     }
 
     /**
@@ -35,7 +102,7 @@ class CouponSeller
     /**
      * @param int $couponId
      * 
-     * @return bool false если купон активен
+     * @return bool false если купон активен и ни разу не использован
      */
     public function deleteCoupon(int $couponId): bool
     {
@@ -55,7 +122,7 @@ class CouponSeller
         }
 
         $coupon = $soldCoupon->getCoupon();
-        if ($coupon && $coupon->getActive())
+        if ($coupon && $coupon->getActive() && $coupon->getUseCount() == 0)
         {
             return false;
         }
@@ -65,6 +132,15 @@ class CouponSeller
         $db->startTransaction();
         try 
         {
+            // Порядок удаления имеет значения (из за обработчика события)
+
+            $result = $soldCoupon->delete();
+            if (!$result->isSuccess())
+            {
+                $db->rollbackTransaction();
+                return false;
+            }
+
             if ($coupon)
             {
                 $result = $coupon->delete();
@@ -74,13 +150,6 @@ class CouponSeller
                     $db->rollbackTransaction();
                     return false;
                 }
-            }
-
-            $result = $soldCoupon->delete();
-            if (!$result->isSuccess())
-            {
-                $db->rollbackTransaction();
-                return false;
             }
         }
         catch (\Exception $exception)
